backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
opencv-python==4.8.1.78
pytesseract==0.3.10
pdf2image==1.16.3
numpy==1.24.3
aiofiles==23.2.1
pydantic==2.5.0

# Optional: For better MIME type detection (Windows users should use python-magic-bin)
# python-magic-bin==0.4.14  # Windows
# python-magic==0.4.27      # Linux/macOS